user showen don and

+++ bistro-boss-client-home-page/src/pages/Dashboard/AllUsers/Allusers.jsx
@@ -1,28 +1,54 @@
 import { FaTrash } from "react-icons/fa";
 import useUsers from "../../../hooks/useUsers";
 import Swal from "sweetalert2";
+import { Helmet } from "react-helmet-async";
+import { useContext } from "react";
+import { Authcontext } from "../../../Providers/AuthProvider";
 
 const Allusers = () => {
   const [users, loading, refetch] = useUsers();
-  const handleDelete = (id) => {
-    fetch(`http://localhost:5000/carts/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        refetch();
-        Swal.fire(
-          "successful!",
-          "this item successfully removed from the cart",
-          "success"
-        );
-        // console.log(cartItems);
-      });
+  const { user, deleteAUser } = useContext(Authcontext);
+  const handleDelete = (id, uid) => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        // deleteAUser(uid)
+        //   .then(() => {
+        // console.log("Successfully deleted user");
+        fetch(`http://localhost:5000/user/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            console.log(data);
+            refetch();
+            Swal.fire(
+              "successful!",
+              "this item successfully deleted the user",
+              "success"
+            );
+            // console.log(cartItems);
+          });
+        // })
+        // .catch((error) => {
+        //   console.log("Error deleting user:", error);
+        // });
+      }
+    });
   };
   return (
     <div>
-      {console.log(users)}
+      {console.log(user)}
+      <Helmet>
+        <title>Bistro Boss || admin || all user</title>
+      </Helmet>
       <div className="pt-10 pe-5 md:pe-10 text-white">
         <div className="bg-gray-700 py-4 px-2 rounded">
           <div className="title uppercase gap-3 justify-between rounded flex">
@@ -84,7 +110,7 @@ const Allusers = () => {
                         <img
                           src={cartItem?.photoURL}
                           alt={cartItem?.displayName}
-                          className="w-16 h-16 object-cover"
+                          className="w-10 h-10 object-cover"
                         />
                       </td>
                       <td className="px-6 py-4 whitespace-nowrap">
@@ -97,11 +123,14 @@ const Allusers = () => {
                       <td className="px-6 py-4 whitespace-nowrap">user</td>
                       <td className="px-6 py-4 whitespace-nowrap">
                         <button
-                          onClick={() => handleDelete(cartItem._id)}
+                          onClick={() =>
+                            handleDelete(cartItem._id, cartItem.uid)
+                          }
                           className="px-4 py-2 text-sm font-medium text-white bg-red-600 rounded-md"
                         >
                           <FaTrash />
                         </button>
+                        {console.log(cartItem.uid)}
                       </td>
                     </tr>
                   ))}
